BookingFoodAPI/requirements.txt
asgiref
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
Deprecated==1.2.13
Django~=5.1.4
django-ckeditor~=6.7.2
django-cors-headers==3.14.0         #install
django-js-asset==2.0.0
django-oauth-toolkit==2.2.0         #install
djangorestframework~=3.15.2
drf-yasg==1.21.5                    #install
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
jwcrypto==1.4.2
MarkupSafe==2.1.2
oauthlib==3.2.2
openapi==1.1.0
packaging==23.0
Pillow==11.1.0                       #install
pycparser==2.21
PyMySQL==1.0.2                      #install
pyrsistent==0.19.3
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.15.0

cloudinary~=1.41.0